+++ include/tcp/asio/tcp_socket.hpp
@@ -354,7 +354,8 @@ public:
 					return;
 				} else {
 
-					if ( pkgLen > readLen) { // exceed buffer size, but user has consume the buffer data
+					if ( pkgLen > (int32_t)readLen) { 
+						// exceed buffer size, but user has consume the buffer data
 						need_package_length = pkgLen - readLen;
 						//read_buffer_pos -= readLen;
 						read_buffer_pos = 0;
